+++ routes/web.php
@@ -1,10 +1,10 @@
 <?php
 
+use App\Http\Controllers\ListingController;
+use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Models\Listing;
-use App\Http\Controllers\ListingController;
-use App\Http\Controllers\UserController;
 
 /*
 |--------------------------------------------------------------------------
@@ -17,29 +17,50 @@ use App\Http\Controllers\UserController;
 |
 */
 
-//All listings
+// Common Resource Routes:
+// index - Show all listings
+// show - Show single listing
+// create - Show form to create new listing
+// store - Store new listing
+// edit - Show form to edit listing
+// update - Update listing
+// destroy - Delete listing  
+
+// All Listings
 Route::get('/', [ListingController::class, 'index']);
 
-//show create post
-Route::get("/listings/create", [ListingController::class, "create"]);
+// Show Create Form
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
-//store listing data
-Route::post("/listings", [ListingController::class, "store"]);
+// Store Listing Data
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
-//show edit form
-Route::get("/listings/{listing}/edit", [ListingController::class, "edit"]);
+// Show Edit Form
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
-//update listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+// Update Listing
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
-//delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+// Delete Listing
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
-//show register form
-Route::get('/register', [UserController::class, 'create']);
+// Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
-//show login form
-Route::get('/login', [UserController::class, 'login']);
+// Single Listing
+Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
-//Single listing
-Route::get("/listing/{listing}", [ListingController::class, 'show']);
+// Show Register/Create Form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+// Create New User
+Route::post('/users', [UserController::class, 'store']);
+
+// Log User Out
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+// Show Login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// Log In User
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
